Fix sliding of waveform when drawing sample in reverse (#7063)

+++ include/Sample.h
@@ -96,7 +96,7 @@ public:
 	auto sampleDuration() const -> std::chrono::milliseconds;
 	auto sampleFile() const -> const QString& { return m_buffer->audioFile(); }
 	auto sampleRate() const -> int { return m_buffer->sampleRate(); }
-	auto sampleSize() const -> int { return m_buffer->size(); }
+	auto sampleSize() const -> size_t { return m_buffer->size(); }
 
 	auto toBase64() const -> QString { return m_buffer->toBase64(); }
